+++ src/test/java/org/trostheide/svgtoolbox/processors/SimplifyProcessorTest.java
@@ -24,7 +24,7 @@ class SimplifyProcessorTest {
         Config config = new Config(
                 "in", "out", 1.0f, Collections.emptyList(), false,
                 HatchStyle.of(45, 5), Collections.emptyMap(), Collections.emptyMap(), Collections.emptyList(), Collections.emptyList(), 0.0,
-                1.0, "linear", 0.0, false, null, false // Tolerance, Pattern, Rotation, Stats, Crop, Optimize
+                1.0, "linear", 45.0, 5.0, 0.0, false, null, false // Tolerance, Pattern, Rotation, Stats, Crop, Optimize
         );
 
         SimplifyProcessor processor = new SimplifyProcessor();
@@ -58,7 +58,7 @@ class SimplifyProcessorTest {
         Config config = new Config(
                 "in", "out", 1.0f, Collections.emptyList(), false,
                 HatchStyle.of(45, 5), Collections.emptyMap(), Collections.emptyMap(), Collections.emptyList(), Collections.emptyList(), 0.0,
-                1.0, "linear", 0.0, false, null, false // Tolerance, Pattern, Rotation, Stats, Crop, Optimize
+                1.0, "linear", 45.0, 5.0, 0.0, false, null, false // Tolerance, Pattern, Rotation, Stats, Crop, Optimize
         );
 
         SimplifyProcessor processor = new SimplifyProcessor();